requirements.txt
# Medical VLM Training Environment Requirements

# Core ML Frameworks
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.25.0
deepspeed>=0.12.0

# Qwen2.5-VL Dependencies
qwen-vl-utils
tiktoken
Pillow>=9.4.0
requests

# Training and Evaluation
datasets>=2.14.0
evaluate>=0.4.0
scikit-learn>=1.3.0
tensorboard>=2.14.0
wandb>=0.16.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Medical Domain Specific
pydicom>=2.4.0
SimpleITK>=2.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
jsonlines>=3.1.0
fire>=0.5.0

# Development and Debugging
ipython>=8.12.0
jupyter>=1.0.0
rich>=13.4.0

# Optional: Efficient Training
flash-attn>=2.3.0
xformers>=0.0.22